created separate email file for badge
